test: share a task_bug_priority scenario across shared-enum tests
The two-model `Task`/`Bug` setup sharing one `Priority { Low, Medium, High }`
embed enum was duplicated in shared_enum_crud and shared_enum_filter.
Extract it into a task_bug_priority scenario.
