+++ crates/toasty-driver-integration-suite/src/scenarios.rs
@@ -1,3 +1,4 @@
+pub mod company_office_address;
 pub mod composite_chain_relations;
 pub mod composite_fk_has_many_belongs_to;
 pub mod composite_has_many_belongs_to;
@@ -16,6 +17,7 @@ pub mod has_many_unique_title;
 pub mod has_one_optional_belongs_to;
 pub mod in_list_item;
 pub mod tagged_item;
+pub mod task_bug_priority;
 pub mod task_with_status;
 pub mod task_with_string_status;
 pub mod two_models;
@@ -33,8 +35,10 @@ pub mod user_two_children;
 pub mod user_unique_email;
 pub mod user_unique_email_with_name;
 pub mod user_with_active;
+pub mod user_with_address;
 pub mod user_with_age;
 pub mod user_with_status;
+pub mod user_with_zip_address;
 pub mod versioned_item;
 pub mod versioned_user_unique_email;
 pub mod widget_mixed_types;
